requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
certifi==2023.5.7
charset-normalizer==3.2.0
Django==4.2.2
django-bootstrap5==23.3
django-filter==23.2
idna==3.4
p5py==1.0.0
pep517==0.13.0
Pillow==8.3.2
psycopg2==2.9.6
requests==2.31.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.0.3
